+++ CHANGELOG.md
@@ -61,6 +61,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `isValid()` to `PGS_ShapePredicates`. Checks if a PShape is valid, and reports the validation error if it is invalid.
 * `obstaclePack()` to `PGS_CirclePacking`. Packs circles of varying radii within a given shape, whilst respecting pointal obstacles.
 * `align()` to `PGS_Transformation`. Aligns one polygon shape to another, by finding the optimal transformation.
+* `extractInnerEdges()` to `PGS_Meshing`. Extracts all inner edges from a mesh.
 
 ### Changed
 * Reimplemented `PGS_Processing.equalParition()`. New algorithm is ~2x faster. Also removed `precise` parameter from method signature (no longer necessary).
@@ -94,6 +95,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Positive-valued offset arguments passed to `point[s]OnExterior()` methods could incorrectly produce offsets towards the interior of a shape. Such values will now always correspond to offset **away** from a shape's interior.
 * `PGS_ShapePredicates.holes()` now identifies and counts gaps in meshes as holes.
 * Quads made by `splitQuadrangulation` were unclosed and are now closed polygons.
+* `PGS_Conversion.toGraph()` no longer adds a spurious closing edges on LINE shapes.
 ### Removed
 *